python/Python-operate/monitor/requirements.txt
appdirs==1.4.4
dukpy==0.2.3
future==0.18.2
javascripthon==0.11
Jinja2==2.11.2
jupyter-echarts-pypkg==0.1.2
lml==0.0.2
macropy3==1.1.0b2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
Pillow==7.1.2
prettytable==0.7.2
protobuf==3.12.2
psutil==5.7.0
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyecharts-snapshot==0.2.0
pyee==7.0.2
pyppeteer==0.2.2
simplejson==3.17.0
six==1.15.0
sockjs-tornado==1.0.7
SQLAlchemy==1.3.17
tornado==6.0.4
tqdm==4.46.1
urllib3==1.25.9
websockets==8.1
